Create minesweeper_expansion.py
This is a program that expands the minesweeper board to reveal neighbouring cells when a non-bomb cell has been selected by the user.

The fully built minesweeper board is given as an input(look at the "minesweeper_code.py" file to understand how to build a minesweeper board when the bomb locations are given). The user shall have to provide the location of the cell they would like to select.
If the choosen cell has a bomb or a non-zero number, the board is returned as it is.  If the choosen cell was a zero cell, we convert it to the value: -2, and expand from it to reveal all the zero cells connected to it.